move ibc util to testutil package
