+++ protocol/testutil/ibc/ibc.go
@@ -1,4 +1,4 @@
-package lib
+package ibc
 
 import (
 	"fmt"
